move desktop page header into desktop
